==> notebooks/2_Hurricane_Paths.ipynb <==
# Analyze Hurricane Paths
### Background
Using our cleaned HURDAT data, we will plot and compare the paths of hurricanes based on different time periods, to get a visual sense of how hurricanes have changed over time due to climate change.
### Contents
1. [Load Data](#1.-Load-Data)
2. [Compare Paths](#2.-Compare-Hurricane-Paths)

import os
import re
import folium
import pandas as pd
import numpy as np
import datetime
import pickle
from urllib.request import urlretrieve
from tqdm import tqdm_notebook as tqdm
from mpl_toolkits.basemap import Basemap

## 1. Load Data

with open('../data/hurricanes_cleaned.pkl', 'rb') as handle:
    hurricanes = pickle.load(handle)

hurricanes['AL062014']

## 2. Compare Hurricane Paths 
Before diving into tesitng our hypotheses, we want to look at some maps to visualize how hurricanes have changed over the last 100 years. We will start by looking at the periods 1918-1927 vs 2008-2017. 

Our hypothesis states that we expect:
1. More hurricanes in the 2008-2017 period
2. A higher number of very strong storms (Cat 4 & 5)

def get_specific_years(hurricane_data, year_start, year_end):
    """
    Takes an input of the JSON data with hurricane ID as key
    Returns an array of dicts for the range (year_end-1).
    """
    year_range = range(year_start, year_end)

    filtered_data = [storm_data for storm_id, storm_data in hurricane_data.items() if storm_data['year'] in year_range]
    
    return filtered_data

def create_path_map(storm_list, category_colors):
    """
    Given a list of storms to plot and the mapping of category strength
    to color, this function will create a folium plot of all the input
    storms and return the map as a folium object.
    """
    # Initialize map
    hurr_map = folium.Map(location=[30, -70], zoom_start = 4)

    # Plot each storm's path with info in tooltip
    for storm in storm_list:
        storm_name = storm['name']

        for idx in range(len(storm['datapoints'])):
            point = storm['datapoints'][idx]

            # Handle last datapoint which will not have a future datapoint
            if idx+1 < len(storm['datapoints']):
                # Current point
                day_points = [(point['lat'], point['lon']), 
                              (storm['datapoints'][idx+1]['lat'], storm['datapoints'][idx+1]['lon'])]

                hurr_info = f"""
                Storm:        {storm_name}
                Date:         {point['record_date']}
                Time:         {point['time']}
                Status:       {point['storm_status']}
                Category:     {point['category']}
                Max Wind:     {point['max_wind']}mph
                Min Pressure: {point['min_pressure']}millibars
                """
                # Plot line
                folium.PolyLine(day_points, 
                                tooltip=storm_name, 
                                popup=hurr_info, 
                                color=category_colors[point['category']],
                               ).add_to(hurr_map)
                
    return hurr_map

# hurr_1908_1917 = get_specific_years(hurricanes, 1908, 1918)
hurr_1908_1917 = get_specific_years(hurricanes, 1960, 1970)

hurr_2008_2017 = get_specific_years(hurricanes, 2008, 2018)

# Define mapping for colors depending on category
category_colors = {1: '#ffb3b3',
                   2: '#ff8080',
                   3: '#e60000',
                   4: '#ff0000',
                   5: '#580808',
                   'TS': '#ffe6e6'}

map_1908_1917 = create_path_map(hurr_1908_1917, category_colors)
map_2008_2017 = create_path_map(hurr_2008_2017, category_colors)

# Save as html
map_1908_1917.save('../maps/hurricane_paths_1908-1917.html')
map_1908_1917.save('../../hm9464.github.io/maps/hurricane_paths_1908-1917.html')
map_2008_2017.save('../maps/hurricane_paths_2008-2017.html')
map_2008_2017.save('../../hm9464.github.io/maps/hurricane_paths_2008-2017.html')
