==> notebooks/3_Hypothesis_Analysis.ipynb <==
# Hypothesis Analysis

import pandas as pd
import numpy as np

## 1. Load Data

# Load hurdat data formatted into dataframe
hurricanes_df = pd.read_csv("../data/hurricanes_df_cleaned.csv")

## 2. Analysis of Hypothesis 1
#### Storms are making more frequent landfalls due to increased strength and duration of systems.
Analysis of:
* Number of storms
* Strong storms (hurricane status)
* Number of landfalls
* Average duration of storms

# % of storms to make landfall for each 10 year period from 1917-2017
periods = {0: {'name': '1918-1927',
               'start': 1918,
               'end': 1927},
           1: {'name': '1928-1937',
               'start': 1928,
               'end': 1937},
           2: {'name': '1938-1947',
               'start': 1938,
               'end': 1947},
           3: {'name': '1948-1957',
               'start': 1948,
               'end': 1957},
           4: {'name': '1958-1967',
               'start': 1958,
               'end': 1967},
           5: {'name': '1968-1977',
               'start': 1968,
               'end': 1977},
           6: {'name': '1978-1987',
               'start': 1978,
               'end': 1987},
           7: {'name': '1988-1997',
               'start': 1988,
               'end': 1997},
           8: {'name': '1998-2007',
               'start': 1998,
               'end': 2007},
           9: {'name': '2008-2017',
               'start': 2018,
               'end': 2017}
          }

hurr_1900s = hurricanes_df[(hurricanes_df['year']>=1918) & (hurricanes_df['year']<=1927)]

#### Number of storms

storm_ids = hurr_1900s['id'].unique()
num_storms = len(storm_ids)
num_storms

#### Average duration of storms

for storm in storm_ids:
    hurr_1900s

### TODO: 
* Plot all cat5 hurricanes in the last 10 years
* Pct of storms that made landfall (https://gis.stackexchange.com/questions/235133/checking-if-a-geocoordinate-point-is-land-or-ocean)
* Filter by storm for each state hit
